+++ homeassistant/components/zwave/__init__.py
@@ -13,8 +13,6 @@ from pprint import pprint
 
 import voluptuous as vol
 
-from homeassistant.components.http import HomeAssistantView
-from homeassistant.const import HTTP_NOT_FOUND
 from homeassistant.core import CoreState
 from homeassistant.loader import get_platform
 from homeassistant.helpers import discovery
@@ -24,13 +22,13 @@ from homeassistant.const import (
 from homeassistant.helpers.entity_values import EntityValues
 from homeassistant.helpers.event import track_time_change
 from homeassistant.util import convert, slugify
-import homeassistant.core as ha
 import homeassistant.config as conf_util
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.dispatcher import (
     async_dispatcher_connect, async_dispatcher_send)
 from homeassistant.components.frontend import register_built_in_panel
 
+from . import api
 from . import const
 from .const import DOMAIN
 from .node_entity import ZWaveBaseEntity, ZWaveNodeEntity
@@ -388,7 +386,7 @@ def setup(hass, config):
     def rename_node(service):
         """Rename a node."""
         node_id = service.data.get(const.ATTR_NODE_ID)
-        node = hass.data[ZWAVE_NETWORK].nodes[node_id]
+        node = network.nodes[node_id]
         name = service.data.get(const.ATTR_NAME)
         node.name = name
         _LOGGER.info(
@@ -506,7 +504,7 @@ def setup(hass, config):
         # to be ready.
         for i in range(const.NETWORK_READY_WAIT_SECS):
             _LOGGER.debug(
-                "network state: %d %s", hass.data[ZWAVE_NETWORK].state,
+                "network state: %d %s", network.state,
                 network.state_str)
             if network.state >= network.STATE_AWAKED:
                 _LOGGER.info("Z-Wave ready after %d seconds", i)
@@ -612,113 +610,13 @@ def setup(hass, config):
 
     if 'frontend' in hass.config.components:
         register_built_in_panel(hass, 'zwave', 'Z-Wave', 'mdi:nfc')
-        hass.http.register_view(ZWaveNodeGroupView)
-        hass.http.register_view(ZWaveNodeConfigView)
-        hass.http.register_view(ZWaveUserCodeView)
+        hass.http.register_view(api.ZWaveNodeGroupView)
+        hass.http.register_view(api.ZWaveNodeConfigView)
+        hass.http.register_view(api.ZWaveUserCodeView)
         hass.http.register_static_path(
             URL_API_OZW_LOG, hass.config.path(OZW_LOG_FILENAME), False)
 
     return True
-
-
-class ZWaveNodeGroupView(HomeAssistantView):
-    """View to return the nodes group configuration."""
-
-    url = "/api/zwave/groups/{node_id}"
-    name = "api:zwave:groups"
-
-    @ha.callback
-    def get(self, request, node_id):
-        """Retrieve groups of node."""
-        from openzwave.group import ZWaveGroup
-
-        hass = request.app['hass']
-        network = hass.data.get(ZWAVE_NETWORK)
-        _LOGGER.info(network.nodes.get(int(node_id)))
-        node = network.nodes.get(int(node_id))
-        if node is None:
-            return self.json_message('Node not found', HTTP_NOT_FOUND)
-        groupdata = node.groups_to_dict()
-        groups = {}
-        for key in groupdata.keys():
-            groupnode = ZWaveGroup(key, network, int(node_id))
-            groups[key] = {'associations': groupnode.associations,
-                           'association_instances':
-                           groupnode.associations_instances,
-                           'label': groupnode.label,
-                           'max_associations': groupnode.max_associations}
-        _LOGGER.info('Groups: %s', groups)
-        if groups:
-            return self.json(groups)
-        else:
-            return self.json_message('Node not found', HTTP_NOT_FOUND)
-
-
-class ZWaveNodeConfigView(HomeAssistantView):
-    """View to return the nodes configuration options."""
-
-    url = "/api/zwave/config/{node_id}"
-    name = "api:zwave:config"
-
-    @ha.callback
-    def get(self, request, node_id):
-        """Retrieve configurations of node."""
-        hass = request.app['hass']
-        network = hass.data.get(ZWAVE_NETWORK)
-        node = network.nodes.get(int(node_id))
-        if node is None:
-            return self.json_message('Node not found', HTTP_NOT_FOUND)
-        config = {}
-        for value in (
-                node.get_values(class_id=const.COMMAND_CLASS_CONFIGURATION)
-                .values()):
-            config[value.index] = {'label': value.label,
-                                   'type': value.type,
-                                   'help': value.help,
-                                   'data_items': value.data_items,
-                                   'data': value.data,
-                                   'max': value.max,
-                                   'min': value.min}
-        _LOGGER.info('Config: %s', config)
-        if config:
-            return self.json(config)
-        else:
-            return self.json_message('Node not found', HTTP_NOT_FOUND)
-
-
-class ZWaveUserCodeView(HomeAssistantView):
-    """View to return the nodes usercode configuration."""
-
-    url = "/api/zwave/usercodes/{node_id}"
-    name = "api:zwave:usercodes"
-
-    @ha.callback
-    def get(self, request, node_id):
-        """Retrieve usercodes of node."""
-        hass = request.app['hass']
-        network = hass.data.get(ZWAVE_NETWORK)
-        node = network.nodes.get(int(node_id))
-        if node is None:
-            return self.json_message('Node not found', HTTP_NOT_FOUND)
-        usercodes = {}
-        if node.has_command_class(const.COMMAND_CLASS_USER_CODE):
-            for value in (
-                    node.get_values(class_id=const.COMMAND_CLASS_USER_CODE)
-                    .values()):
-                if value.genre != const.GENRE_USER:
-                    continue
-                usercodes[value.index] = {'code': value.data,
-                                          'label': value.label,
-                                          'length': len(value.data)}
-            _LOGGER.info('Usercodes: %s', usercodes)
-            if usercodes:
-                return self.json(usercodes)
-            else:
-                return self.json_message('Node does not have usercodes',
-                                         HTTP_NOT_FOUND)
-        else:
-            return self.json_message('Node does not have usercodes',
-                                     HTTP_NOT_FOUND)
 
 
 class ZWaveDeviceEntityValues():
